KOKU-976: KoKu ei tue IE9:iä - (Salon kaupungilla osassa koneissa käytössä IE9)
This commmit implements IE9 -support to Ajanvaraus -form. Implementation caused a bug in searching and adding the recipients for the meeting, this bug is also fixed in this commit.
Issue in Arcusys internal Jira: KOKULT-1: IE9 - päivitys Intalioon / KOKULT-69: Ajanvaraus -form doesn't receive the recipients when converted compatible to IE9.