KOKULT-240: QA TIVA suostumus - sending consent request without selection of the template has layout issues, search for user doesn't work
This task is from Arcusys Jira and is not part of development of new features. This task is for fixing a bug
that occured during development.

Search for users fixed. There was a variable reference error in javascript at commitCustomAutoRowSession
function. Also there was error at username split. Character used for splitting was incorret "/". Fixed with
correct character "\".