KOKULT-277: Send button in wrong place in IE9
When Intalio creates buttons for sending, replacing, canceling etc for Intalio Form it also creates divs that
are supposed to seperate different stacks of buttons from each other. In custom styles there was defined
display: inline for the divs. This definement forced the div and all the buttons after it to the same line.
In this fix definement was removed and it also added new definement that hides the div (display: none).
Updated all koku forms style sheets.