Revert "KOKU-1355: IE9 Fixed buttons moving in all forms except one"
Reverting the jQuery button moving fix that was applied to almost all the forms,
it was causing problems with chrome and firefox and a better solution was found.

By deleting the float:left property that is applied to the intalio startbutton
the moving behaviour is fixed without introducing problems with layout or
buttons in other browsers, plus it's no longer a massive hack.

This reverts commits:
	f6a37d175dec7105bca7915e317676bb26c720fd
	71a4430d8d29c58e235834206f179ff2c1d5dde8
	af7000efe09ba7b0e8d5632564d8fc099e4b670c
	c7d86d019210ebeab009c83753770223cad4e6a0