KOKU-1355: Fixed three bugs in Uusi Pyynto form
KOKULT-690
The default text for the request form that is rendered dynamically to the page
was 'qweqwe'. Changed that to an empty string.

KOKULT-691
The free form question ('Vapaavalintaiset vaihtoehdot') answer adding button
('Lisää vaihtoehto') was slightly in the wrong place (few pixels below the text
area to it's left on the line). Fixed this.

KOKULT-692
IE9 was displaying very small scroll bars for the buttons in the form, disabled
those.