KOKU-1355: Consent on behalf form actions weren't working
KOKULT-762
In consent on behalf of customer form ('Kirjaa asiakkaan suostumus') only some
of the actions ('Toimenpiteet') were being sent to services. This was because of
the clearing the cache of the invisible matrixes (that are supposedly being used
to store the information to be sent to services later on in the process) was
being done inside the for loop that was adding the values to begin with. In
other words only the last bit of information was being sent forward.
Fixed the issue.